requirements.txt
attrs==19.3.0
beancount==2.2.1
beautifulsoup4==4.9.1
bottle==0.12.18
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
lxml==4.5.1
more-itertools==8.4.0
packaging==20.4
plaid-python==8.0.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.5
protobuf==3.12.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
